chore: make into macro
